pyproject.toml
[project]
name = "judge0"
version = "0.0.7"
description = "The official Python SDK for Judge0."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Judge0", email = "contact@judge0.com" }]
classifiers = [
    "Intended Audience :: Developers",

    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",

    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = ["requests>=2.28.0,<3.0.0", "pydantic>=2.0.0,<3.0.0"]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/judge0/judge0-python"
Repository = "https://github.com/judge0/judge0-python.git"
Issues = "https://github.com/judge0/judge0-python/issues"

[dependency-groups]
test = [
    "ufmt==2.7.3",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "pytest-cov==6.0.0",
    "flake8-docstrings==1.7.0",
]
docs = [
    "sphinx==8.1.3",
    "sphinxawesome-theme==5.3.2",
    "sphinx-autodoc-typehints==2.3.0",
    "sphinx-multiversion==0.2.4",
]
dev = [{ include-group = "test" }, { include-group = "docs" }]

[tool.flake8]
extend-ignore = [
    'D100',
    'D101',
    'D102',
    'D103',
    'D104',
    'D105',
    'D107',
    'D205',
    "D209",
    'D400',
    'F821',
]
docstring-convention = "numpy"
max-line-length = 88

[tool.pytest.ini_options]
addopts = "-vv"
